Changes
Changed int to uint8_t
Changed RGB_STRIP pins to STRIP_X_PIN
Added to PINS_DEBUG file

+++ Marlin/RGB_Strip.h
@@ -26,6 +26,6 @@
 
 #include "Arduino.h"
  
-void handle_led_print_event(int code);
+void handle_led_print_event(uint8_t code);
  
 #endif
